QSV: enc_qsv: improve DTS generation code.
If B-pyramid is used, API must be >= 1.6 so we have access to MSDK's DecodeTimeStamp. We can't use it w/VFR, but we can use it to compute the B-frame delay (in number of frames), so we don't have to guess it (which could result in too high a delay being used).



git-svn-id: svn://svn.handbrake.fr/HandBrake/branches/qsv@5633 b64f7644-9d1e-0410-96f1-a4d463321fa5